+++ src/src/Field/VersionField.php
@@ -1,13 +1,6 @@
 <?php
-/*
-GHSVS 2019-02-01
-Usage:
-<field name="assetsbe" type="plgSystemHyphenateGhsvs.assetsbe" hidden="true"
-	loadjs="false" loadcss="true" />
+namespace GHSVS\Plugin\System\ImportfontsGhsvs\Field;
 
-If attributs loadjs or loadcss are missing their default value is TRUE => Assets will be loaded.
-
-*/
 defined('_JEXEC') or die;
 
 use Joomla\CMS\Factory;
@@ -15,9 +8,9 @@ use Joomla\CMS\Form\FormField;
 use Joomla\CMS\Language\Text;
 use Exception;
 
-class plgSystemImportFontsGhsvsFormFieldVersion extends FormField
+class VersionField extends FormField
 {
-	protected $type = 'version';
+	protected $type = 'Version';
 
 	protected function getInput()
 	{
